requirements.txt
aiohttp==3.13.2
fastapi==0.116.1
starlette>=0.41.0
geopy==2.4.1
Jinja2==3.1.6
pillow==11.3.0
psycopg2-binary==2.9.11
pydantic==2.11.9
email-validator>=2.0.0  # requis pour Pydantic EmailStr (validation emails)
PyJWT==2.10.1
python-multipart==0.0.20
requests==2.32.5
resend==2.19.0
sendgrid==6.12.4
stripe==14.0.1
supabase==2.18.1
uvicorn==0.35.0
bcrypt==4.2.1
slowapi==0.1.9
pytest>=7.0.0
python-dotenv>=1.0.0
# Optionnel : sentry-sdk[fastapi] pour monitoring (SENTRY_DSN)
